Refactor: Map Reorg -NO
Non-Operational  Divides the PlasmaPanel into multiple smaller classes
to better organize and modularize the architecture.
