+++ src/main/java/com/csm/rover/simulator/map/modifiers/MapModifier.java
@@ -1,0 +1,9 @@
+package com.csm.rover.simulator.map.modifiers;
+
+import com.csm.rover.simulator.objects.ArrayGrid;
+
+public interface MapModifier {
+
+    void modifyMap(ArrayGrid<Float> map);
+
+}
